requirements.txt
requests>=2.32.0
pandas>=2.0.0
python-dateutil>=2.9.0
tqdm>=4.65.0
